Re #23520

+++ Framework/Algorithms/src/DirectILLTubeBackground.cpp
@@ -1,3 +1,9 @@
+// Mantid Repository : https://github.com/mantidproject/mantid
+//
+// Copyright &copy; 2018 ISIS Rutherford Appleton Laboratory UKRI,
+//     NScD Oak Ridge National Laboratory, European Spallation Source
+//     & Institut Laue - Langevin
+// SPDX - License - Identifier: GPL - 3.0 +
 #include "MantidAlgorithms/DirectILLTubeBackground.h"
 
 #include "MantidAPI/ITableWorkspace.h"
@@ -297,7 +303,7 @@ void DirectILLTubeBackground::exec() {
   std::vector<std::string> const componentNames = components(*instrument);
   API::Progress progress(this, 0.0, 1.0, componentNames.size());
   PARALLEL_FOR_IF(Kernel::threadSafe(*ws))
-  for (int64_t i = 0; static_cast<size_t>(i) < componentNames.size(); ++i) {
+  for (int64_t i = 0; i < static_cast<int64_t>(componentNames.size()); ++i) {
     PARALLEL_START_INTERUPT_REGION
     auto const &componentName = componentNames[static_cast<size_t>(i)];
     progress.report("Processing " + componentName);
